FS-7436: added FEC support for mod_opus
You need an enabled jitter buffer for FEC to work.
If a packet is missing we look in the jitter buffer for the next one,
if the next packet is present we pass it to the decoder to extract
the FEC info from it. If there's no FEC inside the packet, the Opus decoder
will return PLC.