Making mod_rtmp compatible with Adobe Media Server
Adobe Media Server connects differently to mod_rtmp than the way
Flash player connects. The RTMP publish handler rtmp_i_publish
message needs to match the RTMP specification, and a new initStream
handler is required.

This patch modifies the rtmp_i_publish handler to send an onStatus
message to Adobe Media Server that includes an object with "level",
"code" and "description" fields.

The initStream message is sent by Adobe Media Server to notify Freeswitch
of the stream ID for the publish stream. This cannot clash with the play
stream ID so the initStream handler can simply increment the next_streamid
field. The initStream message is undocumented in the RTMP specification.

The transaction ID for onStatus messages has been modified to 0 instead
of 1 to match with the RTMP specification.

FS-7924 #resolve