FS-6954: Use channel flags to check for proxy media or bypass media
Use channel flags CF_PROXY_MODE or CF_PROXY_MEDIA to determine if to disable
SOA when passing a T38 ReINVITE across legs when in proxy media or bypass media.
These channel flags are set when the sofia profile params inbound-bypass-media or
inbound-proxy-media are enabled as well as when the channel variables bypass_media or
proxy_media are set which should handle a few edge cases better.

FS-6954 #resolve.