[GHA] Add pkgconf to macos.yml
Tested on empty Mac Intel and Apple Silicon. Both failed due to missing pkgconf. It is required and not auto downloaded as prerequisite to other packages.

Note the name is now pkgconf on homebrew. It is the newer version of pkg-config and pkgconfig which are now aliases that point to pkgconf on homebrew.

Co-authored-by: MarioG-X <20360699+MarioG-X@users.noreply.github.com>